refactor(cli): extract dogfood command family into cli/commands/dogfood.ts
