+++ apps/axctl/src/cli/index.ts
@@ -123,6 +123,7 @@ import { contextCommand, contextRuntime } from "./commands/context.ts";
 import { projectCommand, projectRuntime } from "./commands/project.ts";
 import { serveCommand, mcpCommand, tuiCommand, serveRuntime } from "./commands/serve.ts";
 import { shareCommand, shareRuntime } from "./commands/share.ts";
+import { dogfoodCommand, dogfoodRuntime } from "./commands/dogfood.ts";
 import type { RuntimeManifest } from "./commands/manifest.ts";
 import { resolvePwdRepository } from "../pwd.ts";
 import { detectStaleness } from "@ax/lib/transcript-staleness";
@@ -136,7 +137,6 @@ import {
 } from "./progress.ts";
 import { AX_VERSION, liveVersionDeps, printVersion, updateAxctl } from "./version.ts";
 import { wantsJson, catchDbErrorAndExit } from "./output.ts";
-import { cmdDogfoodTerminal } from "../dogfood/wterm.ts";
 import {
     buildFileContextHookResponse,
     parseFileContextHookFlags,
@@ -4507,76 +4507,6 @@ const pricingCommand = Command.make(
         }),
 ).pipe(Command.withDescription("Inspect imported model pricing rows"));
 
-const dogfoodTerminalCommand = Command.make(
-    "terminal",
-    {
-        scenario: Flag.choice("scenario", ["axctl-setup", "interactive"] as const).pipe(Flag.withDefault("axctl-setup")),
-        transport: Flag.choice("transport", ["auto", "pty", "process"] as const).pipe(Flag.withDefault("auto")),
-        agent: Flag.choice("agent", ["shell", "claude", "codex", "opencode"] as const).pipe(Flag.optional),
-        command: Flag.string("command").pipe(Flag.optional),
-        successMarker: Flag.string("success-marker").pipe(Flag.optional),
-        timeout: Flag.integer("timeout").pipe(Flag.optional),
-        port: Flag.integer("port").pipe(Flag.withDefault(1742)),
-        json: jsonFlag,
-    },
-    ({ scenario, transport, agent, command, successMarker, timeout, port, json }) =>
-        Effect.promise(() =>
-            cmdDogfoodTerminal([
-                `--scenario=${scenario}`,
-                `--transport=${transport}`,
-                ...stringArg("agent", optionValue(agent)),
-                ...stringArg("command", optionValue(command)),
-                ...stringArg("success-marker", optionValue(successMarker)),
-                ...(timeout._tag === "Some" ? [`--timeout=${timeout.value}`] : []),
-                `--port=${port}`,
-                ...boolArg("json", json),
-            ]),
-        ),
-).pipe(Command.withDescription("Serve a wterm browser terminal dogfood scenario"));
-
-const cmdDogfoodRuns = (args: string[]) =>
-    Effect.gen(function* () {
-        const json = args.includes("--json");
-        const limit = parsePositiveIntFlag("dogfood runs", "limit", args, 30);
-        const db = yield* SurrealClient;
-        const rows = yield* db.query<[Array<Record<string, unknown>>]>(
-            `SELECT id, run_id, scenario, driver, status, agent, command, transport,
-                marker_found, timed_out, timeout_seconds,
-                type::string(started_at) AS started_at,
-                type::string(ended_at) AS ended_at
-            FROM dogfood_run
-            ORDER BY ended_at DESC
-            LIMIT ${limit};`,
-        );
-        const list = rows?.[0] ?? [];
-        if (json) { console.log(prettyPrint(list)); return; }
-        if (list.length === 0) {
-            console.log("(no dogfood runs persisted yet)");
-            return;
-        }
-        for (const row of list) {
-            console.log(
-                `${String(row.ended_at ?? "?")}  [${String(row.status ?? "?")}]  ` +
-                `${String(row.scenario ?? "?")}  ${String(row.driver ?? "?")}  ` +
-                `run_id=${String(row.run_id ?? "?")}`,
-            );
-        }
-    });
-
-const dogfoodRunsCommand = Command.make(
-    "runs",
-    {
-        limit: positiveLimit(30),
-        json: jsonFlag,
-    },
-    ({ limit, json }) => cmdDogfoodRuns([`--limit=${limit}`, ...boolArg("json", json)]),
-).pipe(Command.withDescription("List recent dogfood scenario runs (passed/failed/error)"));
-
-const dogfoodCommand = Command.make("dogfood").pipe(
-    Command.withDescription("Run local dogfood harnesses"),
-    Command.withSubcommands([dogfoodTerminalCommand, dogfoodRunsCommand]),
-);
-
 const searchCommand = Command.make(
     "search",
     {
@@ -5365,7 +5295,6 @@ const LEGACY_RUNTIME: RuntimeManifest = {
     hook: "db",
     hooks: "db",
     agents: "db",
-    dogfood: "db",
 };
 
 export const RUNTIME_BY_COMMAND: RuntimeManifest = {
@@ -5377,6 +5306,7 @@ export const RUNTIME_BY_COMMAND: RuntimeManifest = {
     ...projectRuntime,
     ...serveRuntime,
     ...shareRuntime,
+    ...dogfoodRuntime,
 };
 
 // Commands whose handlers reach into SurrealClient via AppLayer (or the
